refactor: add business logic separation for the ReviewBooking component
