+++ src/firebase.ts
@@ -1,4 +1,6 @@
-import * as firebase from 'firebase';
+import firebase from 'firebase/app'
+import 'firebase/auth'
+import 'firebase/firestore'
 
 let config = null
